% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitscm.R
\name{run_gauche}
\alias{run_gauche}
\title{Create a new Gauche subprocess}
\usage{
run_gauche(command = Sys.which("gosh"), arguments = "-i")
}
\arguments{
\item{command}{path to gosh command}

\item{arguments}{optional arguments}
}
\description{
Create a new Gauche subprocess
}
